requirements.txt
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
influxdb-client>=1.38.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6